refactor(drivers): simpler drivers organization
Closes RNDSTROPPY-104
