+++ pkg/driver/picodata/dialect.go
@@ -1,4 +1,4 @@
-package queries
+package picodata
 
 import (
 	"errors"
