+++ pkg/driver/picodata/dialect_test.go
@@ -1,4 +1,4 @@
-package queries
+package picodata
 
 import (
 	"testing"
