+++ pkg/driver/postgres/dialect.go
@@ -1,4 +1,4 @@
-package queries
+package postgres
 
 import (
 	"errors"
